+++ src/images/images.controller.ts
@@ -1,4 +1,18 @@
-import { Body, Controller, Delete, Get, Param, Query, Post, Put, UploadedFile, UploadedFiles, UseInterceptors } from '@nestjs/common';
+import { 
+    Body,
+    Controller,
+    Delete,
+    Get,
+    Param,
+    Query,
+    Post,
+    Put,
+    UploadedFile,
+    UploadedFiles,
+    UseInterceptors,
+    HttpException, 
+    HttpStatus,
+} from '@nestjs/common';
 import { FileInterceptor, FilesInterceptor } from '@nestjs/platform-express'
 import { ImagesService } from './images.service';
 import { ImagesDTO } from './dtos/images.dto'
@@ -9,17 +23,32 @@ export class ImagesController {
     }
     @Get(':id')
     get(@Param('id') id: string,) {
+
+        if (!id) {
+            throw new HttpException('Missing ID', HttpStatus.BAD_REQUEST);
+        }
+
         return this.service.findById(id);
     }
 
     @Get('album/:id')
     getAllByAlbum(@Param('id') id:string, @Query('page') page: number) {
+
+        if (!id) {
+            throw new HttpException('Missing ID', HttpStatus.BAD_REQUEST);
+        }
+
         return this.service.findAllByAlbum(id, page);
     }
 
     @Post()
     @UseInterceptors(FileInterceptor('file'))
     createOne(@Body() image: ImagesDTO, @UploadedFile() file: Express.Multer.File) {
+        
+        if(!file){
+            throw new HttpException('Missing File', HttpStatus.FORBIDDEN);
+        }
+
         return this.service.create(image, file);
     }
 
@@ -32,11 +61,21 @@ export class ImagesController {
     
     @Put(':id')
     update(@Param('id') id: string, @Body() image: ImagesDTO) {
+        
+        if (!id) {
+            throw new HttpException('Missing ID', HttpStatus.BAD_REQUEST);
+        }
+
         return this.service.update(id, image);
     }
 
     @Delete(':id')
     remove(@Param('id') id: string) {
+        
+        if (!id) {
+            throw new HttpException('Missing ID', HttpStatus.BAD_REQUEST);
+        }
+
         return this.service.remove(id);
     }
